change builderSize to an int (size in GB)

+++ genqlient.go
@@ -362,8 +362,8 @@ type EnsureDepotRemoteBuilderInput struct {
 	AppName *string `json:"appName"`
 	// The scope of the builder; either "app" or "organization"
 	BuilderScope *string `json:"builderScope"`
-	// The size of the builder cache in bytes. Defaults to 50GB
-	BuilderSize *int64 `json:"builderSize"`
+	// The size of the builder cache in GB. Defaults to 50
+	BuilderSize *int `json:"builderSize"`
 	// A unique identifier for the client performing the mutation.
 	ClientMutationId *string `json:"clientMutationId"`
 	// The node ID of the organization
@@ -381,7 +381,7 @@ func (v *EnsureDepotRemoteBuilderInput) GetAppName() *string { return v.AppName 
 func (v *EnsureDepotRemoteBuilderInput) GetBuilderScope() *string { return v.BuilderScope }
 
 // GetBuilderSize returns EnsureDepotRemoteBuilderInput.BuilderSize, and is useful for accessing the field via an interface.
-func (v *EnsureDepotRemoteBuilderInput) GetBuilderSize() *int64 { return v.BuilderSize }
+func (v *EnsureDepotRemoteBuilderInput) GetBuilderSize() *int { return v.BuilderSize }
 
 // GetClientMutationId returns EnsureDepotRemoteBuilderInput.ClientMutationId, and is useful for accessing the field via an interface.
 func (v *EnsureDepotRemoteBuilderInput) GetClientMutationId() *string { return v.ClientMutationId }
